Move resources to Core project.

+++ FuGetGalleryCore/Data/PackagesSearchResults.cs
@@ -47,7 +47,7 @@ namespace FuGetGallery
                         Version = (string)x["version"],
                         Description = (string)x["description"],
                         IconUrl = (string)x["iconUrl"],
-                        TotalDownloads = ((int?)x["totalDownloads"]) ?? 0,
+                        TotalDownloads = ((long?)x["totalDownloads"]) ?? 0,
                         Authors = string.Join (", ", x["authors"]),
                     });
                 }
@@ -109,7 +109,7 @@ namespace FuGetGallery
         public string Version { get; set; } = "";
         public string Description { get; set; } = "";
         public string IconUrl { get; set; } = "";
-        public int TotalDownloads { get; set; } = 0;
+        public long TotalDownloads { get; set; } = 0;
         public string Authors { get; set; } = "";
 
         public override string ToString() => PackageId;
